requirements.txt
Pillow
pyqt5
pyqt5-tools